tests/test_MochiKit-Visual.html

<html>
<head>
    <script type="text/javascript" src="../MochiKit/Base.js"></script>
    <script type="text/javascript" src="../MochiKit/Iter.js"></script>
    <script type="text/javascript" src="../MochiKit/DOM.js"></script>
    <script type="text/javascript" src="../MochiKit/Async.js"></script>
    <script type="text/javascript" src="../MochiKit/Style.js"></script>
    <script type="text/javascript" src="../MochiKit/Color.js"></script>
    <script type="text/javascript" src="../MochiKit/Signal.js"></script>
    <script type="text/javascript" src="../MochiKit/New.js"></script>
    <script type="text/javascript" src="../MochiKit/Visual.js"></script>
    <script type="text/javascript" src="SimpleTest/SimpleTest.js"></script>     

    <link rel="stylesheet" type="text/css" href="SimpleTest/test.css">
    <style type="text/css">
    #elt1, #elt2, #ctn1 {
        visibility: hidden;
    }
    #ctn1 {
        height: 2px;
    }
    </style>
</head>
<body>

<div id='elt1'>elt1</div>
<div id='ctn1'><div id='elt2'></div></div>
<pre id="test">
<script type="text/javascript">
try {
    var lock = new DeferredLock();
    cb1 = function() {
        fade("elt1", {afterFinish: function () {
            is(MochiKit.DOM.isVisible('elt1'), false, "fade ok");
            lock.release();
        }});
    };
    lock.acquire().addCallback(cb1);

    cb2 = function() {
        appear("elt1", {afterFinish: function () {
            is(MochiKit.DOM.isVisible('elt1'), true, "appear ok");
            lock.release();
        }});
    };
    lock.acquire().addCallback(cb2);

    cb3 = function() {
        toggle("elt1", "size", {afterFinish: function () {
            is(MochiKit.DOM.isVisible('elt1'), false, "toggle ok");
            lock.release();
        }});
    };
    lock.acquire().addCallback(cb3);

    cb4 = function() {
        toggle("elt1", "size", {afterFinish: function () {
            is(MochiKit.DOM.isVisible('elt1'), true, "toggle reverse ok");
            lock.release();
        }});
    };
    lock.acquire().addCallback(cb4);

    cb5 = function() {
        switchOff("elt1", {afterFinish: function () {
            is(MochiKit.DOM.isVisible('elt1'), false, "switchOff ok");
            lock.release();
        }});
    };
    lock.acquire().addCallback(cb5);

    cb6 = function() {
        slideUp("ctn1", {afterFinish: function () {
            is(MochiKit.DOM.isVisible('ctn1'), false, "slideUp ok");
            lock.release();
        }});
    };
    lock.acquire().addCallback(cb6);

    cb7 = function() {
        slideDown("ctn1", {afterFinish: function () {
            is(MochiKit.DOM.isVisible('ctn1'), true, "slideDown ok");
            lock.release();
        }});
    };
    lock.acquire().addCallback(cb7);

    lock.acquire().addCallback(function () {
        ok(true, "visual suite finished");
        SimpleTest.finish();
    });

} catch (err) {

    var s = "test suite failure!\n";
    var o = {};
    var k = null;
    for (k in err) {
        // ensure unique keys?!
        if (!o[k]) {
            s +=  k + ": " + err[k] + "\n";
            o[k] = err[k];
        }
    }
    ok ( false, s );
    SimpleTest.finish();

}
</script>
</pre>
</body>
</html>
